Real State Sales Analysis
The dataset consists of lists of unique objects of popular portals for the sale of real estate in Russia. More than 310000 objects. The Russian real estate market has relatively short history. In the Soviet era, all properties were state-owned; people only had the right to use them with apartments allocated based on one's place of work. As a result, options for moving were fairly limited. We want to analyze sales data according to date and check out the highest sales data in each month and the growth rate of sales data.